base/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "AI-agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.3",
    "python-dotenv>=1.2.1",
]

[tool.uv.workspace]
members = [
    "search-agent",
]
